IframeContentRenderer: move to 'preview' subfolder
