+++ packages/perseus-editor/src/item-editor.tsx
@@ -4,8 +4,8 @@ import _ from "underscore";
 
 import DeviceFramer from "./components/device-framer";
 import Editor from "./editor";
-import IframeContentRenderer from "./iframe-content-renderer";
 import ItemExtrasEditor from "./item-extras-editor";
+import IframeContentRenderer from "./preview/iframe-content-renderer";
 
 import type {
     APIOptions,
